:sparkles: user객체 재생성
seq라는 변수가 헷갈려서 변경하였습니다.
