--- src/main/java/codesquad/codestagram/Entity/User.java
@@ -1,41 +1,0 @@
-package codesquad.codestagram.Entity;
-
-import jakarta.persistence.*;
-
-@Entity
-@Table(name = "users")
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long userSeq;
-
-    @Column(nullable = false, unique = true)
-    private String id;
-
-    @Column(nullable = false)
-    private String name;
-
-    @Column(nullable = false)
-    private String password;
-
-    @Column(nullable = false, unique = true)
-    private String email;
-
-    protected User() {}
-
-    public User(String id, String name, String password, String email) {
-        this.id = id;
-        this.name = name;
-        this.password = password;
-        this.email = email;
-    }
-
-    public Long getUserSeq() { return userSeq; }
-    public String getId() { return id; }
-    public String getName() { return name; }
-    public String getPassword() { return password; }
-    public String getEmail() { return email; }
-
-    public void setName(String name) { this.name = name; }
-    public void setEmail(String email) { this.email = email; }
-}
